+++ mne/io/base.py
@@ -18,7 +18,7 @@ from scipy import linalg
 from .constants import FIFF
 from .pick import pick_types, channel_type, pick_channels, pick_info
 from .pick import _pick_data_channels, _pick_data_or_ica
-from .meas_info import write_meas_info
+from .meas_info import write_meas_info, anonymize_info
 from .proj import setup_proj, activate_proj, _proj_equal, ProjMixin
 from ..channels.channels import (ContainsMixin, UpdateChannelsMixin,
                                  SetChannelsMixin, InterpolationMixin)
@@ -665,16 +665,17 @@ class _BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
         self._data[sel, start:stop] = value
 
     def anonymize(self):
-        """Anonymize data
+        """Anonymize data.
 
-        This function will remove ``raw.info['subject_info']`` if it exists.
+        This function will remove 'subject_info', 'meas_date', 'file_id',
+        'meas_id' if they exist in ``raw.info``.
 
         Returns
         -------
         raw : instance of Raw
             The raw object. Operates in place.
         """
-        self.info._anonymize()
+        anonymize_info(self.info)
         return self
 
     @verbose